restructure models and util functions
